+++ clevertap-core/src/main/java/com/clevertap/android/sdk/BaseCallbackManager.java
@@ -26,11 +26,9 @@ public abstract class BaseCallbackManager {
 
     public abstract InAppNotificationButtonListener getInAppNotificationButtonListener();
 
-    public abstract PushPrimerButtonListener getPushPrimerButtonListener();
-
     public abstract InAppNotificationListener getInAppNotificationListener();
 
-    public abstract PushPermissionResponseListener getPushPermissionNotificationResponseListener();
+    public abstract PushPermissionResponseListener getPushPermissionResponseListener();
 
     public abstract CTInboxListener getInboxListener();
 
@@ -62,11 +60,9 @@ public abstract class BaseCallbackManager {
     public abstract void setInAppNotificationButtonListener(
             InAppNotificationButtonListener inAppNotificationButtonListener);
 
-    public abstract void setPushPrimerButtonListener(PushPrimerButtonListener pushPrimerButtonListener);
-
     public abstract void setInAppNotificationListener(InAppNotificationListener inAppNotificationListener);
 
-    public abstract void setPushPermissionNotificationResponseListener(PushPermissionResponseListener pushPermissionNotificationResponseListener);
+    public abstract void setPushPermissionResponseListener(PushPermissionResponseListener pushPermissionResponseListener);
 
     public abstract void setInboxListener(CTInboxListener inboxListener);